+++ pkg/infra/config_test.go
@@ -1,17 +1,17 @@
 package infra_test
 
 import (
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/fixtures"
 	"testing"
 
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/defs"
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/infra"
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/storage/testabilities"
 	"github.com/stretchr/testify/require"
 )
 
 func TestCaseInsensitiveEnums(t *testing.T) {
 	// given:
-	t.Setenv("TEST_SERVER_PRIVATE_KEY", testabilities.StorageServerPrivKey)
+	t.Setenv("TEST_SERVER_PRIVATE_KEY", fixtures.StorageServerPrivKey)
 	t.Setenv("TEST_DB_ENGINE", "SQLite")
 	t.Setenv("TEST_BSV_NETWORK", "MAIN")
 	t.Setenv("TEST_LOGGING_LEVEL", "DeBug")
@@ -65,7 +65,7 @@ func TestEnums(t *testing.T) {
 	for name, test := range tests {
 		t.Run(name, func(t *testing.T) {
 			// given:
-			t.Setenv("TEST_SERVER_PRIVATE_KEY", testabilities.StorageServerPrivKey)
+			t.Setenv("TEST_SERVER_PRIVATE_KEY", fixtures.StorageServerPrivKey)
 			t.Setenv(test.envKey, "wrong")
 
 			// when:
