+++ pkg/internal/fixtures/storage_fixtures.go
@@ -1,0 +1,8 @@
+package fixtures
+
+const (
+	StorageServerPrivKey = "8143f5ed6c5b41c3d084d39d49e161d8dde4b50b0685a4e4ac23959d3b8a319b"
+	StorageIdentityKey   = "028f2daab7808b79368d99eef1ebc2d35cdafe3932cafe3d83cf17837af034ec29" // that matches StorageServerPrivKey
+	StorageName          = "test-storage"
+	StorageHandlerName   = "storage_server"
+)
