+++ pkg/storage/internal/methodtests/find_or_insert_user_test.go
@@ -2,9 +2,9 @@ package methodtests_test
 
 import (
 	"context"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/storage/internal/testabilities"
 	"testing"
 
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/storage/testabilities"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
