+++ pkg/storage/internal/methodtests/make_available_test.go
@@ -2,10 +2,11 @@ package methodtests_test
 
 import (
 	"context"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/internal/fixtures"
+	"github.com/4chain-ag/go-wallet-toolbox/pkg/storage/internal/testabilities"
 	"testing"
 
 	"github.com/4chain-ag/go-wallet-toolbox/pkg/defs"
-	"github.com/4chain-ag/go-wallet-toolbox/pkg/storage/testabilities"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
@@ -22,8 +23,8 @@ func TestMakeAvailable(t *testing.T) {
 	// then:
 	require.NoError(t, err)
 
-	assert.Equal(t, testabilities.StorageName, tableSettings.StorageName)
-	assert.Equal(t, testabilities.StorageIdentityKey, tableSettings.StorageIdentityKey)
+	assert.Equal(t, fixtures.StorageName, tableSettings.StorageName)
+	assert.Equal(t, fixtures.StorageIdentityKey, tableSettings.StorageIdentityKey)
 	assert.Equal(t, defs.NetworkTestnet, tableSettings.Chain)
 	assert.Equal(t, 1024, tableSettings.MaxOutputScript)
 }
